need to figure out stuff... no idea why I am getting nan..also gradient clipping seems to be causing some problems possibly?

+++ Assignment 3 (CIFAR Clustering)/My test code/model.py
@@ -51,11 +51,11 @@ class Autoencoder(nn.Module):
         # Input size: [batch, 3, 32, 32]
         # Output size: [batch, 3, 32, 32]
         self.encoder = nn.Sequential(
-            nn.Conv2d(3, 12, 4, stride=2, padding=1),  # [batch, 12, 16, 16]
+            nn.Conv2d(3, 12, 4, stride=2, padding=1, dtype=torch.float),  # [batch, 12, 16, 16]
             nn.ReLU(),
-            nn.Conv2d(12, 24, 4, stride=2, padding=1),  # [batch, 24, 8, 8]
+            nn.Conv2d(12, 24, 4, stride=2, padding=1, dtype=torch.float),  # [batch, 24, 8, 8]
             nn.ReLU(),
-            nn.Conv2d(24, 48, 4, stride=2, padding=1),  # [batch, 48, 4, 4]
+            nn.Conv2d(24, 48, 4, stride=2, padding=1, dtype=torch.float),  # [batch, 48, 4, 4]
             nn.ReLU(),
             nn.Flatten(),
             nn.Linear(48 * 16, 48 * 16),
@@ -63,18 +63,19 @@ class Autoencoder(nn.Module):
         )
 
         self.decoder = nn.Sequential(
-            nn.ConvTranspose2d(48, 24, 4, stride=2, padding=1),  # [batch, 24, 8, 8]
+            nn.ConvTranspose2d(48, 24, 4, stride=2, padding=1, dtype=torch.float),  # [batch, 24, 8, 8]
             nn.ReLU(),
-            nn.ConvTranspose2d(24, 12, 4, stride=2, padding=1),  # [batch, 12, 16, 16]
+            nn.ConvTranspose2d(24, 12, 4, stride=2, padding=1, dtype=torch.float),  # [batch, 12, 16, 16]
             nn.ReLU(),
-            nn.ConvTranspose2d(12, 3, 4, stride=2, padding=1),  # [batch, 3, 32, 32]
+            nn.ConvTranspose2d(12, 3, 4, stride=2, padding=1, dtype=torch.float),  # [batch, 3, 32, 32]
             nn.Sigmoid(),
         )
 
     def forward(self, x):
+        x = x.to(torch.float)
         encoded = self.encoder(x)
-        reshaped_encoded = encoded.view(-1, 48, 4, 4)
-        decoded = self.decoder(reshaped_encoded)
+        reshaped_encoded = encoded.view(-1, 48, 4, 4).to(torch.float)
+        decoded = self.decoder(reshaped_encoded).to(torch.float)
         return encoded, decoded
 
 
@@ -128,7 +129,6 @@ class RCCPredictor(nn.Module):
         kde_prob_distributions = self.kde(x)  # shape (Batch, 8448)
         rcc_logits = self.stack(kde_prob_distributions)  # shape (Batch, 10)
         return rcc_logits
-
 
 
 if __name__ == '__main__':
